backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic>=2.0.0
pydantic-settings>=2.0.0
langgraph
langchain
langchain-core
langchain-openai
langchain-community
langchain-mcp-adapters
aiosqlite
langgraph-checkpoint-sqlite
orjson>=3.9.0
msgpack>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-multipart>=0.0.6
